- rename to adhere to naming conventions - replace spaces with tabs - fix asterisk import
